Restore working cost field (revert Radix-wrapped cost tooltip)
Wrapping the cost input in a shadcn TooltipTrigger collapses the cost
cell in the editor grid in the browser (the field vanishes). Across the
codebase no tooltip wraps a form input — all 111 TooltipTrigger usages
wrap an icon/button span — so input-wrapping is an unsupported pattern
here, not specific to one layout or the linked state.

Revert the three editors to a plain cost field with a native `title`
hint (shown when the line is linked to a supplier quote) so the field is
guaranteed visible. The shadcn styled tooltip on the cost field is still
open; pursuing it needs to resolve why the Radix trigger collapses the
cell. Keeps the stabilized test i18n mock.
