+++ components/accounting/ClientsOrdersView.tsx
@@ -1141,25 +1141,25 @@ const ClientsOrdersView: React.FC<ClientsOrdersViewProps> = ({
                                         })}
                                       </span>
                                     )}
-                                    <Tooltip disabled={!item.supplierQuoteItemId}>
-                                      <TooltipTrigger asChild>
-                                        <span className="flex items-center gap-1">
-                                          <ValidatedNumberInput
-                                            value={unitCost}
-                                            formatDecimals={2}
-                                            onValueChange={handleCostChange}
-                                            disabled={isReadOnly}
-                                            className={compactInputClass}
-                                          />
-                                          <span className="shrink-0 text-[9px] font-medium text-muted-foreground">
-                                            {currency}
-                                          </span>
-                                        </span>
-                                      </TooltipTrigger>
-                                      <TooltipContent>
-                                        {t('sales:clientQuotes.supplierQuoteCostTooltip')}
-                                      </TooltipContent>
-                                    </Tooltip>
+                                    <div
+                                      className="flex items-center gap-1"
+                                      title={
+                                        item.supplierQuoteItemId
+                                          ? t('sales:clientQuotes.supplierQuoteCostTooltip')
+                                          : undefined
+                                      }
+                                    >
+                                      <ValidatedNumberInput
+                                        value={unitCost}
+                                        formatDecimals={2}
+                                        onValueChange={handleCostChange}
+                                        disabled={isReadOnly}
+                                        className={compactInputClass}
+                                      />
+                                      <span className="shrink-0 text-[9px] font-medium text-muted-foreground">
+                                        {currency}
+                                      </span>
+                                    </div>
                                   </div>
                                 </div>
                                 <div className="space-y-1 lg:col-span-1 lg:space-y-0">
