+++ components/sales/ClientQuotesView.tsx
@@ -1688,25 +1688,18 @@ const ClientQuotesView: React.FC<ClientQuotesViewProps> = ({
                                     statusLabel={statusLabel}
                                   />
                                 </div>
-                                <Tooltip disabled={!isLinkedToSupplierQuote}>
-                                  <TooltipTrigger asChild>
-                                    <span className="flex items-center gap-1">
-                                      <ValidatedNumberInput
-                                        value={cost}
-                                        formatDecimals={2}
-                                        onValueChange={handleCostChange}
-                                        disabled={isReadOnly || isLinkedToSupplierQuote}
-                                        className="w-full text-sm p-2 bg-white border border-zinc-200 rounded-lg focus:ring-1 focus:ring-praetor outline-none text-center disabled:opacity-50 disabled:cursor-not-allowed"
-                                      />
-                                      <span className="text-[9px] font-semibold text-zinc-400 shrink-0">
-                                        {currency}
-                                      </span>
-                                    </span>
-                                  </TooltipTrigger>
-                                  <TooltipContent>
-                                    {t('sales:clientQuotes.supplierQuoteCostTooltip')}
-                                  </TooltipContent>
-                                </Tooltip>
+                                <div className="flex items-center gap-1">
+                                  <ValidatedNumberInput
+                                    value={cost}
+                                    formatDecimals={2}
+                                    onValueChange={handleCostChange}
+                                    disabled={isReadOnly || isLinkedToSupplierQuote}
+                                    className="w-full text-sm p-2 bg-white border border-zinc-200 rounded-lg focus:ring-1 focus:ring-praetor outline-none text-center disabled:opacity-50 disabled:cursor-not-allowed"
+                                  />
+                                  <span className="text-[9px] font-semibold text-zinc-400 shrink-0">
+                                    {currency}
+                                  </span>
+                                </div>
                               </div>
                               <div className="rounded-lg border border-zinc-200 bg-white px-3 py-2 space-y-1">
                                 <div className="text-[10px] font-black text-zinc-400 uppercase tracking-wider flex items-center gap-1">
@@ -1840,25 +1833,25 @@ const ClientQuotesView: React.FC<ClientQuotesViewProps> = ({
                                   </div>
                                 </div>
                                 <div className="col-span-1 flex flex-col items-center justify-center gap-1">
-                                  <Tooltip disabled={!isLinkedToSupplierQuote}>
-                                    <TooltipTrigger asChild>
-                                      <span className="flex items-center gap-1 w-full">
-                                        <ValidatedNumberInput
-                                          value={cost}
-                                          formatDecimals={2}
-                                          onValueChange={handleCostChange}
-                                          disabled={isReadOnly || isLinkedToSupplierQuote}
-                                          className="w-full text-sm px-1 py-2 bg-white border border-zinc-200 rounded-lg focus:ring-1 focus:ring-praetor outline-none text-center disabled:opacity-50 disabled:cursor-not-allowed"
-                                        />
-                                        <span className="text-[9px] font-semibold text-zinc-400 shrink-0">
-                                          {currency}
-                                        </span>
-                                      </span>
-                                    </TooltipTrigger>
-                                    <TooltipContent>
-                                      {t('sales:clientQuotes.supplierQuoteCostTooltip')}
-                                    </TooltipContent>
-                                  </Tooltip>
+                                  <div
+                                    className="flex items-center gap-1 w-full"
+                                    title={
+                                      isLinkedToSupplierQuote
+                                        ? t('sales:clientQuotes.supplierQuoteCostTooltip')
+                                        : undefined
+                                    }
+                                  >
+                                    <ValidatedNumberInput
+                                      value={cost}
+                                      formatDecimals={2}
+                                      onValueChange={handleCostChange}
+                                      disabled={isReadOnly || isLinkedToSupplierQuote}
+                                      className="w-full text-sm px-1 py-2 bg-white border border-zinc-200 rounded-lg focus:ring-1 focus:ring-praetor outline-none text-center disabled:opacity-50 disabled:cursor-not-allowed"
+                                    />
+                                    <span className="text-[9px] font-semibold text-zinc-400 shrink-0">
+                                      {currency}
+                                    </span>
+                                  </div>
                                 </div>
                                 <div className="col-span-1 flex items-center justify-center gap-1">
                                   <ValidatedNumberInput
